+++ CHANGELOG.md
@@ -6,6 +6,7 @@ New functionality added in a backwards-compatible manner:
 * [OLMIS-5335](https://openlmis.atlassian.net/browse/OLMIS-5335): Enabled Adjusted Consumption column for Stock Based Requisitions.
 * [OLMIS-5366](https://openlmis.atlassian.net/browse/OLMIS-5366): Enabled Average Consumption column for Stock Based Requisitions.
 * [OLMIS-5395](https://openlmis.atlassian.net/browse/OLMIS-5395): Enabled Maximum Stock Quantity and Calculated Order Quantity columns for Stock Based Requisitions.
+* [OLMIS-5226](https://openlmis.atlassian.net/browse/OLMIS-5226): Added sort component on the Approve screen in Requisitions.
 
 Improvements:
 * [OLMIS-5409](https://openlmis.atlassian.net/browse/OLMIS-5409): Updated ui-components to version 7.0.0.